Move resetting resp. merging edges onFileStateChanged to its redux service
