+++ visualization/app/codeCharta/state/store/fileSettings/edges/edges.reset.ts
@@ -1,0 +1,38 @@
+import { CCFile, Edge } from "../../../../codeCharta.model"
+import { SettingsMerger } from "../../../../util/settingsMerger"
+
+export function getMergedEdges(inputFiles: CCFile[], withUpdatedPath: boolean = false): Edge[] {
+	let edges: Edge[] = []
+
+	if (inputFiles.length == 1) {
+		return inputFiles[0].settings.fileSettings.edges
+	}
+
+	for (let inputFile of inputFiles) {
+		if (inputFile.settings.fileSettings.edges) {
+			for (let oldEdge of inputFile.settings.fileSettings.edges) {
+				let edge: Edge = {
+					fromNodeName: withUpdatedPath
+						? SettingsMerger.getUpdatedPath(inputFile.fileMeta.fileName, oldEdge.fromNodeName)
+						: oldEdge.fromNodeName,
+					toNodeName: withUpdatedPath
+						? SettingsMerger.getUpdatedPath(inputFile.fileMeta.fileName, oldEdge.toNodeName)
+						: oldEdge.toNodeName,
+					attributes: oldEdge.attributes,
+					visible: oldEdge.visible
+				}
+				const equalEdgeItem = edges.find(e => e.fromNodeName == edge.fromNodeName && e.toNodeName == edge.toNodeName)
+
+				if (equalEdgeItem) {
+					for (let key in edge.attributes) {
+						equalEdgeItem.attributes[key] = edge.attributes[key]
+					}
+				} else {
+					edges.push(edge)
+				}
+			}
+		}
+	}
+
+	return edges
+}
