+++ visualization/app/codeCharta/state/store/fileSettings/edges/edges.service.spec.ts
@@ -5,6 +5,7 @@ import { getService, instantiateModule } from "../../../../../../mocks/ng.mockhe
 import { EdgesAction, EdgesActions } from "./edges.actions"
 import { EdgesService } from "./edges.service"
 import { VALID_EDGE, withMockedEventMethods } from "../../../../util/dataMocks"
+import { FileStateService } from "../../../fileState.service"
 
 describe("EdgesService", () => {
 	let edgesService: EdgesService
@@ -35,6 +36,14 @@ describe("EdgesService", () => {
 			rebuildService()
 
 			expect(StoreService.subscribe).toHaveBeenCalledWith($rootScope, edgesService)
+		})
+
+		it("should subscribe to FileStateService", () => {
+			FileStateService.subscribe = jest.fn()
+
+			rebuildService()
+
+			expect(FileStateService.subscribe).toHaveBeenCalledWith($rootScope, edgesService)
 		})
 	})
